+++ source/extensions/lightspeed.trex.app.style/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 ﻿# Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.5.1]
+## Added
+- Added remap skeleton button icon styles
+
 ## [1.5.0]
 ## Added
 - Added CategoriesFrame ScrollingFrame style